+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project are documented in this file.
 
 ## [Unreleased]
 
+- Changes based on Obsidian plugin submission review
+
 ## [1.0.5] - 2026-07-19
 
 Initial release.